requirements.txt
Django==1.7.1
psycopg2==2.5.4
django-compressor==1.4
django-debug-toolbar==1.1
gunicorn==19.1.1
Fabric==1.10.0
djangorestframework==2.4.4
markdown==2.5.1
django-filter==0.8
